my second commit

+++ controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require("../models/contactModel")
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async(req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 //@desc Post contacts
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async(req, res) => {
     console.log(req.body);
     const {name, email, phone} = req.body;
@@ -23,19 +23,24 @@ const createContact = asyncHandler(async(req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     });
     res.status(201).json(contact);
 });
 
 //@desc put contacts
 //@route PUT /api/contacts
-//@access public
+//@access private
 const updateContact = asyncHandler(async(req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact with this id does not exist")
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("You are not authorized to update another user's contact.")
     }
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id,
         req.body,
@@ -47,12 +52,16 @@ const updateContact = asyncHandler(async(req, res) => {
 
 //@desc Delete contacts
 //@route DELETE /api/contacts
-//@access public
+//@access private
 const deleteContact = asyncHandler(async(req, res) => {
     const deletedContact = await Contact.findById(req.params.id);
     if(!deletedContact){
         res.status(404);
         throw new Error("CONTACT NOT FOUND");
+    }
+    if(deletedContact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("You are not authorized to delete another user's contact.")
     }
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(deletedContact);
@@ -60,7 +69,7 @@ const deleteContact = asyncHandler(async(req, res) => {
 
 //@desc Get a particular contact
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContact = asyncHandler(async(req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
